+++ packages/rake-db/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rake-db
 
+## 2.1.5
+
+### Patch Changes
+
+- Updated dependencies [9b8b3d5]
+  - pqb@0.7.2
+
 ## 2.1.4
 
 ### Patch Changes