+++ packages/test-factory/CHANGELOG.md
@@ -1,5 +1,15 @@
 # orchid-orm-test-factory
 
+## 0.1.3
+
+### Patch Changes
+
+- 9b8b3d5: Prefix unique columns with sequences in test factory
+- Updated dependencies [9b8b3d5]
+  - pqb@0.7.2
+  - orchid-orm@1.3.4
+  - orchid-orm-schema-to-zod@0.1.2
+
 ## 0.1.2
 
 ### Patch Changes